Exercise 6.15 solution
