+++ anecdotes-redux-main/src/components/AnecdoteForm.js
@@ -1,8 +1,7 @@
 import React from "react";
-import {appendAnecdote, newAnecdote} from "../reducers/anecdoteReducer";
+import { createAnecdote } from "../reducers/anecdoteReducer";
 import { messageChange, messageReset } from "../reducers/notificationReducer";
 import { useDispatch } from 'react-redux'
-import anecdoteService from '../services/anecdotes'
 
 const AnecdoteForm = () => {
   const dispatch = useDispatch()
@@ -10,11 +9,8 @@ const AnecdoteForm = () => {
   const create = async (event) => {
     event.preventDefault()
     const anecdote = event.target.anecdote.value
-    console.log('create', anecdote)
-    const newAnecdote = await anecdoteService.createNew(anecdote)
-    console.log('new anecdote', newAnecdote)
-    dispatch(appendAnecdote(newAnecdote))
-    dispatch(messageChange(`Anecdote "${newAnecdote.content}" added`))
+    dispatch(createAnecdote(anecdote))
+    dispatch(messageChange(`Anecdote "${anecdote.content}" added`))
     setTimeout(() => {
       dispatch(messageReset())
     }, 5000)
