+++ anecdotes-redux-main/src/index.js
@@ -1,22 +1,10 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import { configureStore } from '@reduxjs/toolkit'
-import anecdoteReducer from "./reducers/anecdoteReducer";
-import notificationReducer from "./reducers/notificationReducer";
-import filterReducer from "./reducers/filterReducer";
 import { Provider } from 'react-redux'
 import App from './App'
+import store from './store'
 
-const store = configureStore({
-  reducer: {
-    anecdotes: anecdoteReducer,
-    filter: filterReducer,
-    notification: notificationReducer
-  }
-})
-
-console.log(store.getState())
-
+// console.log(store.getState())
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <Provider store={store}>
