+++ anecdotes-redux-main/src/store.js
@@ -1,0 +1,15 @@
+import { configureStore } from '@reduxjs/toolkit'
+import anecdoteReducer from "./reducers/anecdoteReducer";
+import notificationReducer from "./reducers/notificationReducer";
+import filterReducer from "./reducers/filterReducer";
+
+
+const store = configureStore({
+  reducer: {
+    anecdotes: anecdoteReducer,
+    filter: filterReducer,
+    notification: notificationReducer
+  }
+})
+
+export default store
